+++ newnipt/server/__init__.py
@@ -1,4 +1,3 @@
-
 import os
 import logging
 
@@ -13,21 +12,22 @@ logging.basicConfig(level=logging.INFO)
 LOG = logging.getLogger(__name__)
 
 
-def create_app(test = False):
+def create_app(test=False):
     app = Flask(__name__, instance_relative_config=True)
     if not test:
         app.config.from_object(f"{__name__}.config")
-        client = MongoClient(app.config['DB_URI'])
-        db_name = app.config['DB_NAME']
+        client = MongoClient(app.config["DB_URI"])
+        db_name = app.config["DB_NAME"]
         app.client = client
         app.db = client[db_name]
-        app.adapter = NiptAdapter(client, db_name = db_name)
-        app.analysis_path = app.config['ANALYSIS_PATH']
+        app.adapter = NiptAdapter(client, db_name=db_name)
+        app.analysis_path = app.config["ANALYSIS_PATH"]
         app.register_blueprint(login_bp)
         app.register_blueprint(server_bp)
-        login_manager.init_app(app)        
+        login_manager.init_app(app)
 
-        if app.config['DEBUG']==1:
+        if app.config["DEBUG"] == 1:
             from flask_debugtoolbar import DebugToolbarExtension
-            toolbar = DebugToolbarExtension(app)        
+
+            toolbar = DebugToolbarExtension(app)
     return app