+++ CHANGELOG.md
@@ -3,5 +3,12 @@
 This change log will document the notable changes to this project in this file and it is following [Semantic Versioning](https://semver.org/)
 
 ## [x.x.x]
+### Added
+- Sample views 
+- Batch views
+
+### Fixed
+
+### Changed